split track-data into two parts
