+++ src/app/main/utility-components/track-data-utility/track-data/track-data-graph/track-data-graph.component.css
@@ -1,0 +1,130 @@
+.track-data {
+  position: relative;
+  overflow: hidden;
+}
+
+header {
+  display: flex;
+  padding-block: var(--spacing-4);
+  gap: var(--spacing-2);
+  justify-content: space-between;
+  padding-inline: var(--spacing-4);
+}
+
+/* CHART STYLES */
+
+.chart-wrapper {
+  position: relative;
+}
+
+.legend {
+  position: absolute;
+  bottom: 125px;
+  right: 100px;
+  display: flex;
+  flex-direction: column;
+  background-color: var(--color-background-base-default);
+  padding: var(--spacing-2);
+}
+
+.legend-pushed {
+  right: 460px;
+  bottom: 145px;
+}
+
+.legend rux-checkbox {
+  padding: var(--spacing-1);
+}
+
+.button-wrapper {
+  display: flex;
+  gap: var(--spacing-2);
+}
+
+.button-wrapper rux-button {
+  width: var(--spacing-24);
+}
+
+.toolbar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding-block: var(--spacing-2);
+  padding-inline: var(--spacing-4);
+  background-color: var(--color-background-surface-header);
+}
+
+.toolbar-buttons {
+  display: flex;
+  align-items: center;
+  gap: var(--spacing-4);
+}
+
+rux-slider {
+  width: 13rem;
+  padding-inline: var(--spacing-2);
+}
+
+.zoom-wrapper {
+  display: flex;
+}
+
+footer {
+  display: flex;
+  justify-content: center;
+  gap: var(--spacing-8);
+  padding-bottom: var(--spacing-6);
+  position: fixed;
+  bottom: 0;
+  text-align: center;
+  width: 82%;
+}
+
+/* NG-DEEP is eventually going to be deprecated, but there is no alternative yet */
+:host ::ng-deep .tooltip-box {
+  background-color: var(--color-background-base-selected);
+  padding-inline: var(--spacing-3);
+  padding-block: var(--spacing-1);
+  border: none;
+}
+
+:host ::ng-deep .selected-data-point {
+  fill: var(--color-data-visualization-1);
+}
+
+/* TABLE STYLES */
+
+.table-wrapper {
+  height: 38rem;
+}
+
+.icon-span {
+  width: var(--spacing-6);
+  display: inline-block;
+  padding-right: 0;
+}
+
+th.show-icon {
+  padding-right: var(--spacing-6);
+}
+
+rux-icon {
+  vertical-align: middle;
+}
+
+th:nth-of-type(3):hover,
+th:nth-of-type(2):hover {
+  cursor: pointer;
+}
+
+td:nth-of-type(3),
+td:nth-of-type(2),
+th:nth-of-type(3),
+th:nth-of-type(2) {
+  text-align: right;
+}
+
+td:nth-of-type(3),
+td:nth-of-type(2) {
+  padding-right: var(--spacing-8);
+}
