+++ src/app/main/utility-components/track-data-utility/track-data/track-data-graph/track-data-graph.component.html
@@ -1,0 +1,58 @@
+<div class="toolbar">
+  <div class="toolbar-buttons">
+    <rux-button
+      [disabled]="disableUndo"
+      (click)="onUndo()"
+      iconOnly
+      icon="undo"
+    />
+    <rux-button (click)="onDelete()" iconOnly icon="delete" />
+  </div>
+
+  <div class="zoom-wrapper">
+    <rux-icon icon="search" size="extra-small" />
+    <rux-slider
+      [value]="zoomLevel"
+      (ruxchange)="handleZoom($event)"
+      min="5"
+      max="20"
+    />
+    <rux-icon icon="search" size="1.5rem" />
+  </div>
+</div>
+
+<div class="chart-wrapper">
+  <apx-chart
+    *ngIf="series"
+    [series]="series"
+    [chart]="chartOptions.chart"
+    [xaxis]="chartOptions.xaxis"
+    [yaxis]="chartOptions.yaxis"
+    [markers]="chartOptions.markers"
+    [stroke]="chartOptions.stroke"
+    [legend]="chartOptions.legend"
+    [tooltip]="chartOptions.tooltip"
+  ></apx-chart>
+
+  <div class="legend" #legend>
+    <rux-checkbox
+      disabled
+      label="Inc"
+      (ruxchange)="filterCheckboxes($event, 'Inc')"
+    />
+    <rux-checkbox
+      disabled
+      label="El"
+      (ruxchange)="filterCheckboxes($event, 'El')"
+    />
+    <rux-checkbox
+      disabled
+      label="Range"
+      (ruxchange)="filterCheckboxes($event, 'Range')"
+    />
+  </div>
+</div>
+<footer>
+  <span>OD Quality Est: 82/100</span>
+  <span>Data Points: {{ dataPointLength }}</span>
+</footer>
