+++ src/app/main/utility-components/track-data-utility/track-data/track-data-table/track-data-table.component.spec.ts
@@ -1,0 +1,21 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { TrackDataTableComponent } from './track-data-table.component';
+
+describe('TrackDataComponent', () => {
+  let component: TrackDataTableComponent;
+  let fixture: ComponentFixture<TrackDataTableComponent>;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [TrackDataTableComponent],
+    });
+    fixture = TestBed.createComponent(TrackDataTableComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
