+++ src/app/main/utility-components/track-data-utility/track-data/track-data.component.html
@@ -28,14 +28,14 @@
         size="small"
         secondary
         [disabled]="isSettingsDrawerOpen"
-        (click)="openSitesDrawer()"
+        (click)="toggleSitesDrawer()"
         >Sites</rux-button
       >
       <rux-button
         *ngIf="isSitesDrawerOpen"
         size="small"
         secondary
-        (click)="openSitesDrawer()"
+        (click)="toggleSitesDrawer()"
         icon="keyboard-backspace"
         >Close</rux-button
       >
@@ -44,121 +44,25 @@
         size="small"
         secondary
         [disabled]="isSitesDrawerOpen"
-        (click)="openSettingsDrawer()"
+        (click)="toggleSettingsDrawer()"
         >Settings</rux-button
       >
       <rux-button
         *ngIf="isSettingsDrawerOpen"
         size="small"
         secondary
-        (click)="openSettingsDrawer()"
+        (click)="toggleSettingsDrawer()"
         icon="keyboard-backspace"
         >Close</rux-button
       >
     </div>
   </header>
   <div *ngIf="showGraph">
-    <div class="toolbar">
-      <div class="toolbar-buttons">
-        <rux-button
-          [disabled]="disableUndo"
-          (click)="onUndo()"
-          iconOnly
-          icon="undo"
-        />
-        <rux-button (click)="onDelete()" iconOnly icon="delete" />
-      </div>
-
-      <div class="zoom-wrapper">
-        <rux-icon icon="search" size="extra-small" />
-        <rux-slider
-          [value]="zoomLevel"
-          (ruxchange)="handleZoom($event)"
-          min="5"
-          max="20"
-        />
-        <rux-icon icon="search" size="1.5rem" />
-      </div>
-    </div>
-
-    <div class="chart-wrapper">
-      <apx-chart
-        [series]="series"
-        [chart]="chartOptions.chart"
-        [xaxis]="chartOptions.xaxis"
-        [yaxis]="chartOptions.yaxis"
-        [markers]="chartOptions.markers"
-        [stroke]="chartOptions.stroke"
-        [legend]="chartOptions.legend"
-        [tooltip]="chartOptions.tooltip"
-      ></apx-chart>
-
-      <div class="legend" #legend>
-        <rux-checkbox
-          disabled
-          label="Inc"
-          (ruxchange)="filterCheckboxes($event, 'Inc')"
-        />
-        <rux-checkbox
-          disabled
-          label="El"
-          (ruxchange)="filterCheckboxes($event, 'El')"
-        />
-        <rux-checkbox
-          disabled
-          label="Range"
-          (ruxchange)="filterCheckboxes($event, 'Range')"
-        />
-      </div>
-    </div>
-    <footer>
-      <span>OD Quality Est: 82/100</span>
-      <span>Data Points: {{ dataPointLength }}</span>
-    </footer>
+    <fds-track-data-graph #trackDataGraph></fds-track-data-graph>
   </div>
 
   <!--Table Component: ToDo split into two components-->
   <div *ngIf="showTable">
-    <div class="table-wrapper">
-      <table>
-        <thead>
-          <tr>
-            <th
-              *ngFor="let col of trackFilesTableService.columns"
-              [ngClass]="{ sortable: col.sortable }"
-              [attr.type]="col.type"
-            >
-              <div
-                [ngClass]="{ active: col.sorted }"
-                class="h-stack"
-                (click)="trackFilesTableService.sort(col)"
-              >
-                {{ col.header }}
-                <rux-icon *ngIf="col.sortable" [icon]="col.icon" size="small" />
-              </div>
-            </th>
-          </tr>
-        </thead>
-        <tbody>
-          <ng-container *ngFor="let row of trackFilesTableService.data">
-            <tr *ngIf="!row.filtered">
-              <td
-                *ngFor="let col of trackFilesTableService.columns"
-                [ngSwitch]="col.field"
-                [attr.type]="col.type"
-                [ngClass]="{ sortable: col.sortable }"
-              >
-                <ng-container *ngSwitchCase="'creationDate'">
-                  {{ row.creationDate | date: 'MM/dd/YYYY' }}
-                </ng-container>
-                <ng-container *ngSwitchDefault>
-                  {{ row[col.field] }}
-                </ng-container>
-              </td>
-            </tr>
-          </ng-container>
-        </tbody>
-      </table>
-    </div>
+    <fds-track-data-table></fds-track-data-table>
   </div>
 </div>
